requirements.txt
# Telegram bot
python-telegram-bot>=21.0

# OCR engines (выберите один основной)
pytesseract>=0.3.10
# paddleocr  # опционально, для лучшего качества (требует paddlepaddle)
# easyocr    # опционально, fallback

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.20.0

# Schema & validation
pydantic>=2.0
python-dotenv>=1.0

# Fuzzy matching (OCR error correction)
rapidfuzz>=3.0

# Excel export
openpyxl>=3.1.0

# Async file I/O
aiofiles>=23.0.0

# HTTP (Yandex Vision, etc.)
requests>=2.28.0

# PDF support (опционально)
# pdf2image  # требует poppler
